Change stepper error estimate storage
Aside from cleanup, the main effect of this change is that the
ErrorControl stepper now always uses the most recent error estimate,
where it used to only activate if that estimate was from the most
recent evaluation.  The only case where that matters is for LTS
updating with a multistage method, which used to not work and now uses
the error estimates from the previous two steps, as opposed to the
current and previous step for single-stage LTS.  This matches the
behavior for GTS adjustment, so should be acceptable for now.